Make sure that only an admin can stop the server
Now the "stop" query takes the usual sessionid/token parameters. The
server checks if the session exists, and is attached to a user with
the admin permission. Only then is the server allowed to stop.

Of course that touches all the server tests since they all finish by
a stop query.