Prepare advanced user management
If simply "creating users" may be called "advanced management" :-)

Add a new message: "user" that will be invoked by the administrator
to create or merely modify a user.

Add an "email" field to the user, that will be ultimately used to
send temporary credentials.

Add a "validity" field to the user, that gives a password a max
date, used when creating a temporary password to the user.